Fix build on mac os

+++ test/engine/ValuesForTesting.h
@@ -249,5 +249,5 @@ class ValuesForTestingNoKnownEmptyResult : public ValuesForTesting {
  public:
   using ValuesForTesting::ValuesForTesting;
   bool knownEmptyResult() override { return false; }
-  size_t getSizeEstimateBeforeLimit() override { return 1; }
+  uint64_t getSizeEstimateBeforeLimit() override { return 1; }
 };
